Add getaddrinfo interposer
Host lookups which begin with perf-test-hydra- will resolve to the IP
address of the current machine. But that can also be overriden by
setting PROXY_IP=<IP/ADDR>. I'd ordinarily use LD_PRELOAD to force
this behaviour but I couldn't work out how to make that work on RHEL
with selinux disabled (or enabled).

This is only to facilitate the perforance testing I have in:

  https://github.com/frobware/haproxy-openshift