Add getaddrinfo interposer
Host lookups which begin with "perf-test-hydra" will resolve to the IP
address of the current machine. But that can be overriden by the
environment variable PROXY_IP=<IP/ADDR>.

An alternative to baking this into mb is to turn this into a library
that can then be LD_PRELOADed.

This is only to facilitate the performance testing I have in:

  https://github.com/frobware/haproxy-openshift